+++ core/src/main/java/io/fabric8/maven/core/service/kubernetes/DockerBuildService.java
@@ -20,6 +20,7 @@ import java.util.Objects;
 import io.fabric8.maven.core.service.BuildService;
 import io.fabric8.maven.core.service.Fabric8ServiceException;
 import io.fabric8.maven.docker.config.ImageConfiguration;
+import io.fabric8.maven.docker.service.ImagePullManager;
 import io.fabric8.maven.docker.service.ServiceHub;
 
 /**
@@ -45,8 +46,9 @@ public class DockerBuildService implements BuildService {
 
         io.fabric8.maven.docker.service.BuildService dockerBuildService = dockerServiceHub.getBuildService();
         io.fabric8.maven.docker.service.BuildService.BuildContext dockerBuildContext = config.getDockerBuildContext();
+        ImagePullManager imagePullManager = config.getImagePullManager();
         try {
-            dockerBuildService.buildImage(imageConfig, dockerBuildContext);
+            dockerBuildService.buildImage(imageConfig, imagePullManager, dockerBuildContext);
 
             // Assume we always want to tag
             dockerBuildService.tagImage(imageConfig.getName(), imageConfig);