+++ core/src/main/java/io/fabric8/maven/core/service/openshift/OpenshiftBuildService.java
@@ -341,7 +341,7 @@ public class OpenshiftBuildService implements BuildService {
             fromImage = extractBaseFromConfiguration(buildConfig);
         }
 
-        String pullRegistry = EnvUtil.findRegistry(new ImageName(fromImage).getRegistry(), dockerBuildContext.getPullRegistry(), dockerBuildContext.getRegistryConfig().getRegistry());;
+        String pullRegistry = EnvUtil.fistRegistryOf(new ImageName(fromImage).getRegistry(), dockerBuildContext.getRegistryConfig().getRegistry(), dockerBuildContext.getRegistryConfig().getRegistry());;
 
         if (pullRegistry != null) {
             RegistryService.RegistryConfig registryConfig = dockerBuildContext.getRegistryConfig();
@@ -423,8 +423,8 @@ public class OpenshiftBuildService implements BuildService {
             File fullDockerFilePath = buildConfig.getAbsoluteDockerFilePath(buildContext.getMojoParameters());
             fromImage = DockerFileUtil.extractBaseImage(
                     fullDockerFilePath,
-                    buildContext.getMojoParameters().getProject().getProperties(),
-                    buildConfig.getFilter());
+                    DockerFileUtil.createInterpolator(buildContext.getMojoParameters(),
+                    buildConfig.getFilter()));
         } catch (IOException e) {
             // Cant extract base image, so we wont try an auto pull. An error will occur later anyway when
             // building the image, so we are passive here.